Gestione aggiungi opera in un file a parte #77
